Rename Service Fabric Membership to Clustering
