+++ Samples/ServiceFabric/Stateless/StatelessCalculatorService/StatelessCalculatorService.csproj
@@ -10,7 +10,7 @@
     <PackageReference Include="Microsoft.Orleans.Core.Abstractions" Version="$(OrleansPackageVersion)" />
     <PackageReference Include="Microsoft.Orleans.OrleansCodeGenerator.Build" Version="$(OrleansPackageVersion)" PrivateAssets="All" />
     <PackageReference Include="Microsoft.Orleans.Hosting.ServiceFabric" Version="$(OrleansPackageVersion)" />
-    <PackageReference Include="Microsoft.Orleans.Membership.ServiceFabric" Version="$(OrleansPackageVersion)" />
+    <PackageReference Include="Microsoft.Orleans.Clustering.ServiceFabric" Version="$(OrleansPackageVersion)" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\GrainInterfaces\GrainInterfaces.csproj" />
