+++ src/Orleans.Clustering.ServiceFabric/Models/ServicePartitionSilos.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using Orleans.ServiceFabric;
 
-namespace Orleans.Membership.ServiceFabric
+namespace Orleans.Clustering.ServiceFabric.Models
 {
     /// <summary>
     /// Represents a Service Fabric service partition and the Orleans silos within it.
