+++ src/Orleans.Clustering.ServiceFabric/Orleans.Clustering.ServiceFabric.csproj
@@ -1,15 +1,15 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <PackageId>Microsoft.Orleans.Membership.ServiceFabric</PackageId>
-    <Title>Microsoft Orleans Service Fabric Membership Support</Title>
-    <Description>Support for hosting Microsoft Orleans on Service Fabric.</Description>
+    <PackageId>Microsoft.Orleans.Clustering.ServiceFabric</PackageId>
+    <Title>Microsoft Orleans Service Fabric Clustering Support</Title>
+    <Description>Support for using Service Fabric for cluster information when hosting Microsoft Orleans on Service Fabric.</Description>
     <PackageTags>$(PackageTags) Azure ServiceFabric</PackageTags>
   </PropertyGroup>
 
   <PropertyGroup>
     <TargetFramework>net461</TargetFramework>
-    <AssemblyName>Orleans.Membership.ServiceFabric</AssemblyName>
-    <RootNamespace>Orleans.Membership.ServiceFabric</RootNamespace>
+    <AssemblyName>Orleans.Clustering.ServiceFabric</AssemblyName>
+    <RootNamespace>Orleans.Clustering.ServiceFabric</RootNamespace>
     <PlatformTarget>x64</PlatformTarget>
   </PropertyGroup>
