+++ src/Orleans.Clustering.ServiceFabric/Properties/AssemblyInfo.cs
