+++ src/Orleans.Clustering.ServiceFabric/Utilities/FabricQueryManager.cs
@@ -6,8 +6,9 @@ using System.Fabric.Query;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.ServiceFabric.Services.Client;
+using Orleans.Clustering.ServiceFabric.Models;
 
-namespace Orleans.Membership.ServiceFabric.Utilities
+namespace Orleans.Clustering.ServiceFabric.Utilities
 {
     internal class FabricQueryManager : IFabricQueryManager
     {
