+++ src/Orleans.ServiceFabric/Orleans.ServiceFabric.csproj
@@ -10,6 +10,6 @@
   </PropertyGroup>
   <ItemGroup>
     <ProjectReference Include="..\Orleans.Hosting.ServiceFabric\Orleans.Hosting.ServiceFabric.csproj" />
-    <ProjectReference Include="..\Orleans.Membership.ServiceFabric\Orleans.Membership.ServiceFabric.csproj" />
+    <ProjectReference Include="..\Orleans.Clustering.ServiceFabric\Orleans.Clustering.ServiceFabric.csproj" />
   </ItemGroup>
 </Project>
